gold/scripts/py/03_test_sharepoint_connection.py
#!/usr/bin/env python3
"""
Public Excel to PostgreSQL Data Sync Script

This script reads publicly accessible Excel files (from SharePoint, OneDrive, or web URLs)
and persists the data to a PostgreSQL database. Each Excel tab becomes a separate table.
Works with files that can be accessed without authentication (incognito browser works).
"""

import os
import sys
import logging
import time
import hashlib
import io
from datetime import datetime
from typing import Dict, List, Any, Optional
import pandas as pd
import psycopg2
from psycopg2 import sql
from psycopg2.extras import execute_values
import schedule
import requests
from dotenv import load_dotenv

load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('logs/03_test_sharepoint_connection.log'),
        logging.StreamHandler(sys.stdout)
    ]
)
logger = logging.getLogger(__name__)

# TODO: main implementation to test the connection to SharePoint
