gold/requirements.txt
openpyxl>=3.1.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
pyodbc>=5.0.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
sqlparse>=0.4.4
